moved models folder, added 4 user methods
