+++ src/models/documentNotFoundError.ts
